+++ crates/cli/src/workflows/install_workflow.rs
@@ -650,18 +650,16 @@ impl InstallWorkflowManager {
                         ProgressState::Exit => {
                             break;
                         }
-                        ProgressState::Message(message) => {
-                            if !console.out.is_quiet() {
-                                let _ = console.out.write_line_with_prefix(
-                                    apply_style_tags(
-                                        // Compatibility with the UI theme
-                                        message
-                                            .replace("version>", "hash>")
-                                            .replace("versionalt>", "symbol>"),
-                                    ),
-                                    &prefix,
-                                );
-                            }
+                        ProgressState::Message(message) if !console.out.is_quiet() => {
+                            let _ = console.out.write_line_with_prefix(
+                                apply_style_tags(
+                                    // Compatibility with the UI theme
+                                    message
+                                        .replace("version>", "hash>")
+                                        .replace("versionalt>", "symbol>"),
+                                ),
+                                &prefix,
+                            );
                         }
                         _ => {}
                     }