+++ CHANGELOG.md
@@ -15,6 +15,26 @@
 - [Rust](https://github.com/moonrepo/plugins/blob/master/tools/rust/CHANGELOG.md)
 - [Schema (TOML, JSON, YAML)](https://github.com/moonrepo/plugins/blob/master/tools/internal-schema/CHANGELOG.md)
 
+## Unreleased
+
+#### 💥 Breaking
+
+- Reverted the `proto exec --raw` changes from v0.56 until we can properly handle all edge cases.
+
+#### 🚀 Updates
+
+- Added in-memory locking for plugin downloads with the same URL, to avoid race conditions where multiple threads attempt to download/rename/copy files at the same time. This primarily applies to the `npm`, `pnpm`, and `yarn` plugins, as they all utilize the same underlying WASM plugin.
+
+#### 🐞 Fixes
+
+- Fixed an issue where arguments with spaces or special characters were not quoted correctly when running in a shell, resulting in a syntax parse error.
+- Potential fix for the "Failed to load and create plugin: expected at least one module field" error.
+
+#### ⚙️ Internal
+
+- Updated Rust to v1.95.0.
+- Updated dependencies.
+
 ## 0.56.2
 
 #### 🚀 Updates